+++ tap_heap/exceptions.py
@@ -1,0 +1,2 @@
+class HeapForbiddenError(Exception):
+    """Raised when all streams are inaccessible due to 403 Forbidden errors."""