+++ tests/base.py
@@ -5,6 +5,7 @@ from tap_tester.base_suite_tests.base_case import BaseCase
 
 class TapHeapBaseCase(BaseCase):
     start_date = "2021-04-07T00:00:00Z"
+    IS_FORBIDDEN_STREAM = "is-forbidden-stream"
 
     @staticmethod
     def tap_name():
@@ -33,6 +34,15 @@ class TapHeapBaseCase(BaseCase):
             'pageviews': {
                 BaseCase.PRIMARY_KEYS: {'event_id'}
             }
+        }
+
+    @classmethod
+    def expected_stream_names(cls):
+        """The expected stream names, excluding forbidden streams."""
+        return {
+            stream_name
+            for stream_name, md in cls.expected_metadata().items()
+            if not md.get(cls.IS_FORBIDDEN_STREAM, False)
         }
 
     def get_properties(self):